Support dev env in fabfile

+++ fabfile.py
@@ -1,5 +1,6 @@
 import os
 from os import path
+import json
 
 from fabric.api import env, run, cd, local, lcd
 
@@ -9,26 +10,33 @@ BRANCH = 'master'
 PYTHON = '{0}/venv/bin/python'.format(CODE_DIR)
 PIP = '{0}/venv/bin/pip'.format(CODE_DIR)
 
-def fetch_hosts():
+def _fetch_hosts():
     with lcd(DEPLOY_DIR):
-        output = local('ansible tag_Name_video_processor --list-hosts -i inventory/production', capture=True)
-        return [ip.strip() for ip in output.splitlines()[1:]]
+        output = local('python inventory/production', capture=True)
+        return json.loads(output)
 
-
+HOSTS = _fetch_hosts()
 env.user = 'videalize'
-env.hosts = fetch_hosts()
+env.hosts = HOSTS['tag_Name_video_processor']
 env.key_filename = path.join(os.environ['HOME'], '.ssh', 'videalize')
 env.use_ssh_config = True
 
+def _is_production():
+    return env.host_string in HOSTS['tag_Env_production']
 
 def deploy():
     with cd(CODE_DIR):
         run("GIT_SSH_COMMAND='ssh -i ~/.ssh/video_processor_deploy' git fetch")
-        sha = run("git rev-parse origin/{0}".format(BRANCH))
-        run("git checkout {0}".format(sha))
+        if _is_production():
+            sha = run("git rev-parse origin/{0}".format(BRANCH))
+            run("git checkout {0}".format(sha))
 
-        run("{0} install -r requirements.txt".format(PIP))
-        run("{0} setup.py install".format(PYTHON))
+            run("{0} install -r requirements.txt".format(PIP))
+            run("{0} setup.py install".format(PYTHON))
 
-        pid = run("cat tmp/video_processor.pid")
-        run("kill -SIGINT {0}".format(pid))
+            pid = run("cat tmp/video_processor.pid")
+            run("kill -SIGINT {0}".format(pid))
+        else:
+            run("git checkout {0}".format(BRANCH))
+            run("git merge origin/{0}".format(BRANCH))
+            run("{0} install -r requirements.txt".format(PIP))
